Make ODA prefer components in the current project, instead of unrelated projects in the workspace.
The IndexManager indexes all projects in the workspace equally.  Before
making this change, the search for the declaration of a component would
return the first one in the entire workspace, whether it's in the
current project or not.

This change determines the project from the EditorInput (which is a
FileEditorInput) and passes that to the search function, which then
determined a "project scope" list that lists the current project, along
with any projects it depends on. The component search will restrict the
search to the projects in that list.

This points out that there still could be some ambiguity if the
component is defined in the current project AND in a project it depends
on.  The current search will just return the first one found, most
likely from the alphabetically first project name.  It might be more
reasonable to prefer the current projec, and then any other projects in
the scope list.